requirements.txt
absl-py==0.12.0
aiohttp==3.7.3
aiohttp-cors==0.7.0
aioredis==1.3.1
alabaster==0.7.12
appdirs==1.4.4
astor==0.8.1
astroid==2.5.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autopep8==1.5.5
Babel==2.9.0
backcall==0.2.0
bcrypt==3.2.0
black==20.8b1
bleach==3.3.0
blessings==1.7
cachetools==4.2.1
certifi==2020.12.5
cffi==1.14.5
chardet==3.0.4
click==7.1.2
cloudpickle==1.6.0
colorama==0.4.4
colorful==0.5.4
cryptography==3.4.6
cycler==0.10.0
decorator==4.4.2
defusedxml==0.7.1
diff-match-patch==20200713
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
filelock==3.0.12
flake8==3.8.4
flatbuffers==1.12
gast==0.3.3
google-api-core==1.26.0
google-auth==1.27.0
google-auth-oauthlib==0.4.3
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
grpcio==1.32.0
h5py==2.10.0
helpdev==0.7.1
hiredis==1.1.0
idna==2.10
imagesize==1.2.0
importlib-metadata==3.7.2
intervaltree==3.1.0
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
isort==5.7.0
jedi==0.17.2
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==23.0.0
kiwisolver==1.3.1
lazy-object-proxy==1.5.2
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.4
mccabe==0.6.1
mistune==0.8.4
msgpack==1.0.2
mtcnn==0.1.0
multidict==5.1.0
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.2
nest-asyncio==1.5.1
nltk==3.4.5
numpy==1.19.5
numpydoc==1.1.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencensus==0.7.12
opencensus-context==0.1.2
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
openpyxl==3.0.7
opt-einsum==3.3.0
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
paramiko==2.7.2
parso==0.7.0
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
prometheus-client==0.9.0
prompt-toolkit==3.0.16
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
py-spy==0.3.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.8.1
pylint==2.7.2
pyls-black==0.4.6
pyls-spyder==0.3.2
PyNaCl==1.4.0
pyparsing==2.4.7
PyQt5==5.12.3
PyQt5-sip==12.8.1
PyQtWebEngine==5.12.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-jsonrpc-server==0.4.0
python-language-server==0.36.2
pytz==2021.1
pywin32==300
pywin32-ctypes==0.2.0
PyYAML==5.4.1
pyzmq==22.0.3
QDarkStyle==2.8.1
QtAwesome==1.0.2
qtconsole==5.0.2
QtPy==1.9.0
ray==1.2.0
redis==3.5.3
regex==2020.11.13
requests==2.25.1
requests-oauthlib==1.3.0
rope==0.18.0
rsa==4.7.1
Rtree==0.9.7
scikit-learn==0.24.1
scipy==1.4.1
sentencepiece==0.1.83
six==1.15.0
sklearn==0.0
snowballstemmer==2.1.0
sortedcontainers==2.3.0
Sphinx==3.5.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.2.3
spyder-kernels==1.10.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.2.0
tensorflow-gpu-estimator==2.2.0
termcolor==1.1.0
testpath==0.4.4
textdistance==4.2.1
threadpoolctl==2.1.0
three-merge==0.1.1
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typed-ast==1.4.2
typing-extensions==3.7.4.3
ujson==4.0.2
urllib3==1.26.3
watchdog==1.0.2
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
xlwt==1.3.0
yapf==0.30.0
yarl==1.6.3
zipp==3.4.0
